refactor(causal): refact causal page ui
